add immutable-js library, refactor actions
